ISPN-10322 Remove HotRodCommandInitializer in favour of InitializableCommand
