+++ server/hotrod/src/main/java/org/infinispan/server/hotrod/command/HotRodModuleCommandExtensions.java
@@ -2,12 +2,10 @@ package org.infinispan.server.hotrod.command;
 
 import org.infinispan.commands.module.ModuleCommandExtensions;
 import org.infinispan.commands.module.ModuleCommandFactory;
-import org.infinispan.commands.module.ModuleCommandInitializer;
 import org.infinispan.commands.remote.CacheRpcCommand;
 
 /**
- * It register the {@link HotRodCommandFactory} and {@link HotRodCommandInitializer} to handle the {@link
- * CacheRpcCommand} used by this module.
+ * It register the {@link HotRodCommandFactory} to handle the {@link CacheRpcCommand} used by this module.
  *
  * @author Pedro Ruivo
  * @since 9.1
@@ -17,10 +15,5 @@ public final class HotRodModuleCommandExtensions implements ModuleCommandExtensi
    @Override
    public ModuleCommandFactory getModuleCommandFactory() {
       return new HotRodCommandFactory();
-   }
-
-   @Override
-   public ModuleCommandInitializer getModuleCommandInitializer() {
-      return new HotRodCommandInitializer();
    }
 }
